requirements.txt
certifi==2021.10.8
chardet==3.0.4
Django==2.2.27
django-constance==2.8.0
django-cors-headers==3.0.2
django-extra-fields==1.2.3
django-picklefield==3.0.1
djangorestframework==3.9.4
idna==2.8
mysqlclient==1.4.2.post1
numpy==1.22.2
pandas==1.4.1
pdfkit==0.6.1
Pillow==9.0.0
psycopg2-binary==2.8.6
PyPDF2==1.26.0
python-dateutil==2.8.2
pytz==2021.3
qrcode==6.1
requests==2.22.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.25.11
xlwt==1.3.0
